chore: move Execution Worker to pkg/testworkflows
